refactor: split backend into modules and clean up device flow
Splits the single lib.rs into auth.rs, github.rs, tray.rs, and a
thin lib.rs that just wires everything together.

While here, addresses review findings:

- replace fragile `format!("{:?}", issue.state)` with an explicit
  match on octocrab's IssueState enum
- skip the initial sleep in poll_device_flow so the happy path
  doesn't wait a full interval before the first request
- share a single reqwest::Client via OnceLock
- drop `.text() + from_str` in favor of `.json()`, reading the
  body only on error for better diagnostics
- tighten the mutex scope in fetch_watched
- collapse the mount-time `is_authenticated` round-trip into a
  direct fetch_watched call with silent failure
- rename `set_auto_hide(enabled)` + `suppress_auto_hide` to
  `set_window_pinned(pinned)` + `pinned` to kill the double-negation
- extract DEVICE_FLOW_TIMEOUT constant
- replace `deviceCode!` with `{@const dc}` in the template
- drop the direct `chrono` dependency (only reached via octocrab)
