+++ src-tauri/Cargo.toml
@@ -23,7 +23,6 @@ tauri-plugin-opener = "2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 octocrab = "0.49.7"
-chrono = { version = "0.4.44", features = ["serde"] }
 reqwest = { version = "0.13.2", default-features = false, features = ["rustls", "json", "form"] }
 tokio = { version = "1.52.1", features = ["rt", "time"] }
