+++ src-tauri/src/tray.rs
@@ -1,0 +1,52 @@
+use tauri::{
+    image::Image,
+    menu::{Menu, MenuItem},
+    tray::{MouseButton, MouseButtonState, TrayIconBuilder, TrayIconEvent},
+    App, Manager, PhysicalPosition,
+};
+
+const TRAY_ICON_BYTES: &[u8] = include_bytes!("../icons/tray-icon.png");
+
+pub fn setup(app: &App) -> tauri::Result<()> {
+    let quit_item = MenuItem::with_id(app, "quit", "Quit eir", true, None::<&str>)?;
+    let menu = Menu::with_items(app, &[&quit_item])?;
+
+    let tray_icon = Image::from_bytes(TRAY_ICON_BYTES)?;
+    TrayIconBuilder::with_id("main")
+        .icon(tray_icon)
+        .icon_as_template(true)
+        .menu(&menu)
+        .show_menu_on_left_click(false)
+        .on_menu_event(|app, event| {
+            if event.id.as_ref() == "quit" {
+                app.exit(0);
+            }
+        })
+        .on_tray_icon_event(|tray, event| {
+            if let TrayIconEvent::Click {
+                button: MouseButton::Left,
+                button_state: MouseButtonState::Up,
+                position,
+                ..
+            } = event
+            {
+                let app = tray.app_handle();
+                let Some(window) = app.get_webview_window("main") else {
+                    return;
+                };
+                if window.is_visible().unwrap_or(false) {
+                    let _ = window.hide();
+                    return;
+                }
+                let size = window.outer_size().unwrap_or_default();
+                let x = position.x as i32 - (size.width as i32) / 2;
+                let y = position.y as i32 + 8;
+                let _ = window.set_position(PhysicalPosition { x, y });
+                let _ = window.show();
+                let _ = window.set_focus();
+            }
+        })
+        .build(app)?;
+
+    Ok(())
+}
